refactor: intefaces 패키지 추가 (presentation을 interfaces 패키지 하위로 변경)
