+++ src/main/java/com/sion/concertbooking/intefaces/presentation/response/WaitingQueueInfoResponse.java
@@ -1,4 +1,4 @@
-package com.sion.concertbooking.presentation.response;
+package com.sion.concertbooking.intefaces.presentation.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueDetailInfo;
