+++ src/test/java/com/sion/concertbooking/presentation/rest/PointControllerTest.java
@@ -5,6 +5,7 @@ import com.sion.concertbooking.application.PointCharger;
 import com.sion.concertbooking.application.PointDeductor;
 import com.sion.concertbooking.application.result.PointResult;
 import com.sion.concertbooking.domain.pointhistory.TransactionType;
+import com.sion.concertbooking.intefaces.presentation.rest.PointController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
