+++ src/test/java/com/sion/concertbooking/presentation/rest/ReservationControllerTest.java
@@ -8,7 +8,8 @@ import com.sion.concertbooking.domain.reservation.ReservationStatus;
 import com.sion.concertbooking.domain.seat.SeatGrade;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueStatus;
 import com.sion.concertbooking.infrastructure.aspect.TokenInfo;
-import com.sion.concertbooking.presentation.request.ConcertReservationCreateRequest;
+import com.sion.concertbooking.intefaces.presentation.request.ConcertReservationCreateRequest;
+import com.sion.concertbooking.intefaces.presentation.rest.ReservationController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
