+++ src/test/java/com/sion/concertbooking/presentation/rest/WaitingQueueControllerTest.java
@@ -6,7 +6,8 @@ import com.sion.concertbooking.domain.watingqueue.WaitingQueueDetailInfo;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueStatus;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueService;
 import com.sion.concertbooking.infrastructure.aspect.TokenInfo;
-import com.sion.concertbooking.presentation.request.WaitingQueueRegisterRequest;
+import com.sion.concertbooking.intefaces.presentation.request.WaitingQueueRegisterRequest;
+import com.sion.concertbooking.intefaces.presentation.rest.WaitingQueueController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
